+++ assets/src/js/wp-plugins/woocommerce/woo-ajax-add-to-cart.js
@@ -1,4 +1,3 @@
-import { data } from "infinite-scroll";
 import { options } from "../../constants";
 import delegate from "delegate";
 
@@ -53,23 +52,7 @@ class WooAjaxAddToCart {
 
     const addToCartBtn = event.delegateTarget;
     const form = addToCartBtn.closest("form.cart");
-
-    if (!form.checkValidity()) {
-      form.reportValidity();
-      return;
-    }
-
     const formData = this.#getFormData(form);
-
-    const formDataObj = this.#convertDataToObject(formData)
-
-    if (formDataObj['add-to-cart']) {
-      // Assign the value of 'add-to-cart' to 'product_id'
-      formDataObj['product_id'] = formDataObj['add-to-cart'];
-
-      // Delete the original 'add-to-cart' key
-      delete formDataObj['add-to-cart'];
-    }
 
     if (formData.some(({ name }) => name === "add-to-cart")) {
       event.preventDefault();
@@ -83,7 +66,7 @@ class WooAjaxAddToCart {
        */
       jQuery("body").trigger("adding_to_cart", [
         jQuery(addToCartBtn),
-        formDataObj,
+        formData,
       ]);
 
       /**
@@ -92,15 +75,10 @@ class WooAjaxAddToCart {
        */
       jQuery.ajax({
         type: "POST",
-        url: oceanwpLocalize.wc_ajax_url.toString().replace( '%%endpoint%%', 'add_to_cart' ),
-        data: formDataObj,
+        url: oceanwpLocalize.wc_ajax_url,
+        data: formData,
+
         success: function (response) {
-
-          if ( response.error && response.product_url ) {
-						window.location = response.product_url;
-						return;
-					}
-
           /**
            * Because Woocommerce plugin uses jQuery custom event,
            * We also have to use jQuery to customize this event.
@@ -118,7 +96,6 @@ class WooAjaxAddToCart {
             return;
           }
         },
-        dataType: 'json',
       });
     }
   };
@@ -133,8 +110,7 @@ class WooAjaxAddToCart {
       // Add view cart text.
       if (
         !options.is_cart &&
-        !cartBtn.parentNode.querySelector(".added_to_cart") &&
-        cart_hash
+        !cartBtn.parentNode.querySelector(".added_to_cart")
       ) {
         cartBtn.insertAdjacentHTML(
           "afterend",
@@ -143,12 +119,6 @@ class WooAjaxAddToCart {
       }
     }
   };
-
-  // #getFormData = (form) => {
-  //   form = form instanceof Element ? form : document.querySelector(form);
-  //   return new FormData(form); // Directly return the FormData object
-  // };
-
 
   #getFormData = (form) => {
     form = form instanceof Element ? form : document.querySelector(form);
@@ -185,19 +155,6 @@ class WooAjaxAddToCart {
           };
     });
   };
-
-  #convertDataToObject = (dataArray) => {
-    // Initialize an empty object to hold the converted data
-    const dataObject = {};
-
-    // Iterate over the array of data
-    dataArray.forEach(item => {
-      // Set each item into the object with its name as the key and value as the value
-      dataObject[item.name] = item.value;
-    });
-
-    return dataObject;
-  }
 }
 
 jQuery(function () {