+++ inc/customizer/options/woocommerce.php
@@ -3955,7 +3955,7 @@ $options = [
 										'key' => 'normal',
 										'label' => esc_html__( 'Select Color', 'oceanwp' ),
 										'selector' => [
-											'.woocommerce ul.products li.product .button,.woocommerce ul.products li.product .product-inner .added_to_cart' => 'background-color'
+											'.woocommerce ul.products li.product .button,.woocommerce ul.products li.product .product-inner .added_to_cart, .woocommerce ul.products li.product:not(.product-category) .image-wrap .button' => 'background-color'
 										],
 										'attr' => [
 											'transport' => 'postMessage',
@@ -3966,7 +3966,7 @@ $options = [
 										'key' => 'hover',
 										'label' => esc_html__( 'Hover', 'oceanwp' ),
 										'selector' => [
-											'.woocommerce ul.products li.product .button:hover,.woocommerce ul.products li.product .product-inner .added_to_cart:hover' => 'background-color'
+											'.woocommerce ul.products li.product .button:hover,.woocommerce ul.products li.product .product-inner .added_to_cart:hover, .woocommerce ul.products li.product:not(.product-category) .image-wrap .button:hover' => 'background-color'
 										],
 										'attr' => [
 											'transport' => 'postMessage',
@@ -3991,7 +3991,7 @@ $options = [
 										'key' => 'normal',
 										'label' => esc_html__( 'Select Color', 'oceanwp' ),
 										'selector' => [
-											'.woocommerce ul.products li.product .button,.woocommerce ul.products li.product .product-inner .added_to_cart' => 'color'
+											'.woocommerce ul.products li.product .button,.woocommerce ul.products li.product .product-inner .added_to_cart, .woocommerce ul.products li.product:not(.product-category) .image-wrap .button' => 'color'
 										],
 										'attr' => [
 											'transport' => 'postMessage',
@@ -4003,7 +4003,7 @@ $options = [
 										'key' => 'hover',
 										'label' => esc_html__( 'Hover', 'oceanwp' ),
 										'selector' => [
-											'.woocommerce ul.products li.product .button:hover,.woocommerce ul.products li.product .product-inner .added_to_cart:hover' => 'color'
+											'.woocommerce ul.products li.product .button:hover,.woocommerce ul.products li.product .product-inner .added_to_cart:hover, .woocommerce ul.products li.product:not(.product-category) .image-wrap .button:hover' => 'color'
 										],
 										'attr' => [
 											'transport' => 'postMessage',
@@ -4029,7 +4029,7 @@ $options = [
 										'key' => 'normal',
 										'label' => esc_html__( 'Select Color', 'oceanwp' ),
 										'selector' => [
-											'.woocommerce ul.products li.product .button,.woocommerce ul.products li.product .product-inner .added_to_cart' => 'border-color'
+											'.woocommerce ul.products li.product .button,.woocommerce ul.products li.product .product-inner .added_to_cart, .woocommerce ul.products li.product:not(.product-category) .image-wrap .button' => 'border-color'
 										],
 										'attr' => [
 											'transport' => 'postMessage',
@@ -4041,7 +4041,7 @@ $options = [
 										'key' => 'hover',
 										'label' => esc_html__( 'Hover', 'oceanwp' ),
 										'selector' => [
-											'.woocommerce ul.products li.product .button:hover,.woocommerce ul.products li.product .product-inner .added_to_cart:hover' => 'border-color'
+											'.woocommerce ul.products li.product .button:hover,.woocommerce ul.products li.product .product-inner .added_to_cart:, .woocommerce ul.products li.product:not(.product-category) .image-wrap .button:hover' => 'border-color'
 										],
 										'attr' => [
 											'transport' => 'postMessage',
@@ -6216,7 +6216,7 @@ $options = [
 								'key' => 'normal',
 								'label' => esc_html__( 'Select Color', 'oceanwp' ),
 								'selector' => [
-									'.woocommerce div.product div[itemprop="description"]' => 'color'
+									'.woocommerce div.product div[itemprop="description"], .woocommerce-product-details__short-description' => 'color'
 								],
 								'attr' => [
 									'transport' => 'postMessage',